Move test to own file
